# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Gym Me BackEnd Thesis",
        "title": "Gym Me Backend Thesis",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "dasdasdas",
            "url": "asdasdas",
            "email": "tes@gmail.com"
        },
        "license": {
            "name": "Gymme"
        },
        "version": "1.0"
    },
    "host": "localhost:3000",
    "paths": {
...
        "/api/information": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get All Information By Pagination",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Information"
                ],
                "summary": "Get All Information By Pagination",
                "parameters": [
                    {
                        "type": "string",
                        "description": "sort_by",
                        "name": "sort_by",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "sort_of",
                        "name": "sort_of",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "page",
                        "name": "page",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "limit",
                        "name": "limit",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/entities.InformationEntities"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponses"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponses"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponses"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponses"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponses"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponses"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create New Information",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Information"
                ],
                "summary": "Create New Information",
                "parameters": [
                    {
                        "description": "Insert Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/MenuPayloads.InformationInsertPayloads"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.StandarAPIResponses"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponses"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponses"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponses"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponses"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponses"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponses"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update Information",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Information"
                ],
                "summary": "Update Information",
                "parameters": [
                    {
                        "description": "Update Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/MenuPayloads.InformationUpdatePayloads"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponses"
                        }
                    }
                }
            }
        },
        "/api/information/by-id/{information_id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get Information By Information",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Information"
                ],
                "summary": "Get Information By Information",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "information_id",
                        "name": "information_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entities.InformationEntities"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponses"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponses"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponses"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponses"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponses"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponses"
                        }
                    }
                }
            }
        },
        "/api/information/delete/{information_id}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete Information",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Information"
                ],
                "summary": "Delete Information",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "information_id",
                        "name": "information_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.StandarAPIResponses"
                        }
                    }
                }
            }
        },
        "/api/information/search": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get All Information By Pagination With Filter",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Information"
                ],
                "summary": "Get All Information By Pagination With Filter",
                "parameters": [
                    {
                        "type": "string",
                        "description": "key_filter",
                        "name": "key_filter",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "sort_by",
                        "name": "sort_by",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "sort_of",
                        "name": "sort_of",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "page",
                        "name": "page",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "limit",
                        "name": "limit",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/entities.InformationEntities"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponses"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponses"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponses"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponses"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponses"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponses"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "MenuPayloads.InformationBodyDetail": {
            "type": "object",
            "properties": {
                "information_body_paragraph": {
                    "type": "string"
                },
                "information_image_content_path": {
                    "type": "string"
                }
            }
        },
        "MenuPayloads.InformationInsertPayloads": {
            "type": "object",
            "properties": {
                "information_body_paragraph": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/MenuPayloads.InformationBodyDetail"
                    }
                },
                "information_header": {
                    "type": "string"
                },
                "information_id": {
                    "type": "integer"
                },
                "information_image_content_path_1": {
                    "type": "string"
                },
                "information_image_content_path_2": {
                    "type": "string"
                },
                "information_image_content_path_3": {
                    "type": "string"
                },
                "information_image_content_path_4": {
                    "type": "string"
                },
                "information_image_content_path_5": {
                    "type": "string"
                },
                "information_type_id": {
                    "type": "integer"
                }
            }
        },
...
        "MenuPayloads.InformationUpdatePayloads": {
            "type": "object",
            "properties": {
                "information_body_content": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/MenuPayloads.InformationBodyDetail"
                    }
                },
                "information_id": {
                    "type": "integer"
                }
            }
        },
...
        "entities.InformationBodyEntities": {
            "type": "object",
            "properties": {
                "information_body_id": {
                    "type": "integer"
                },
                "information_body_paragraph": {
                    "type": "string"
                },
                "information_id": {
                    "type": "integer"
                },
                "information_image_content_path": {
                    "type": "string"
                }
            }
        },
        "entities.InformationEntities": {
            "type": "object",
            "properties": {
                "informationType": {
                    "$ref": "#/definitions/entities.InformationType"
                },
                "information_body": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entities.InformationBodyEntities"
                    }
                },
                "information_created_by_user_id": {
                    "type": "integer"
                },
                "information_date_created": {
                    "type": "string"
                },
                "information_header": {
                    "type": "string"
                },
                "information_id": {
                    "type": "integer"
                },
                "information_type_id": {
                    "type": "integer"
                }
            }
        },
        "entities.InformationType": {
            "type": "object",
            "properties": {
                "information_type_id": {
                    "type": "integer"
                },
                "information_type_name": {
                    "type": "string"
                }
            }
        },
...
        "responses.ErrorResponses": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "status_code": {
                    "type": "integer"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "responses.StandarAPIResponses": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "status_code": {
                    "type": "integer"
                },
                "success": {
                    "type": "boolean"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}